feat: moved types and constant imports to ssi-snap-types package
