+++ packages/connector/src/utils.ts
@@ -13,6 +13,7 @@ export type GetSnapsResponse = {
     initialPermissions?: { [k: string]: unknown };
   };
 };
+
 async function getWalletSnaps(): Promise<GetSnapsResponse> {
   return await window.ethereum.request({
     method: 'wallet_getSnaps',
