+++ packages/snap/src/rpc/did/getAvailableMethods.ts
@@ -1,4 +1,4 @@
-import { availableMethods } from '../../constants/index';
+import { availableMethods } from '@blockchain-lab-um/ssi-snap-types';
 
 export function getAvailableMethods(): string[] {
   return availableMethods.map((key) => key);
