+++ packages/snap/src/utils/didUtils.ts
@@ -1,5 +1,5 @@
 import { SnapProvider } from '@metamask/snap-types';
-import { availableMethods, availableVCStores } from '../constants/index';
+import { availableMethods, availableVCStores } from '@blockchain-lab-um/ssi-snap-types';
 import { getDidKeyIdentifier } from '../did/key/keyDidUtils';
 import { SSISnapState } from '../interfaces';
 import { getCurrentNetwork } from './snapUtils';
